requirements.txt
fastapi
uvicorn
python-multipart
pandas
openpyxl
jinja2
reportlab
python-dotenv
google-genai
matplotlib